chore: hacky fixes

+++ scripts/db.sh
@@ -2,29 +2,35 @@
 
 set -euo pipefail
 
-function createPostgresConfig() {
-  cp /etc/postgresql/$PG_VERSION/main/postgresql.custom.conf.tmpl /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
-  sudo -u postgres echo "autovacuum = $AUTOVACUUM" >> /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
-  cat /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
-}
-
-function setPostgresPassword() {
-    sudo -u postgres psql -c "ALTER USER renderer PASSWORD '${PGPASSWORD:-renderer}'"
-}
-
-
 if [ "$1" = "db" ]; then
-    # Clean /tmp
-    rm -rf /tmp/*
 
-    # Fix postgres data privileges
-    chown -R postgres: /var/lib/postgresql/ /data/database/postgres/
+    if [ ! -f /DB_INITIALIZED ]; then
+        # Ensure that database directory is in right state
+        chown postgres:postgres -R /var/lib/postgresql
 
-    # Initialize PostgreSQL and Apache
-    createPostgresConfig
-    service postgresql start
-    # service apache2 restart
-    setPostgresPassword
+        if [ ! -f /var/lib/postgresql/12/main/PG_VERSION ]; then
+            sudo -u postgres /usr/lib/postgresql/12/bin/pg_ctl -D /var/lib/postgresql/12/main/ initdb -o "--locale C.UTF-8"
+        fi
+
+        cp /etc/postgresql/$PG_VERSION/main/postgresql.custom.conf.tmpl /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
+        sudo -u postgres echo "autovacuum = $AUTOVACUUM" >> /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
+        cat /etc/postgresql/$PG_VERSION/main/conf.d/postgresql.custom.conf
+
+        service postgresql start
+
+        sudo -u postgres createuser renderer
+        sudo -u postgres createdb -E UTF8 -O renderer gis
+        sudo -u postgres psql -d gis -c "CREATE EXTENSION postgis;"
+        sudo -u postgres psql -d gis -c "CREATE EXTENSION hstore;"
+        sudo -u postgres psql -d gis -c "ALTER TABLE geometry_columns OWNER TO renderer;"
+        sudo -u postgres psql -d gis -c "ALTER TABLE spatial_ref_sys OWNER TO renderer;"
+        sudo -u postgres psql -c "ALTER USER renderer PASSWORD '${PGPASSWORD:-renderer}'"
+
+        # Fix postgres data privileges
+        chown -R postgres: /var/lib/postgresql/ /data/database/postgres/
+    else
+        service postgresql start
+    fi
 
     # Run while handling docker stop's SIGTERM
     stop_handler() {
@@ -34,15 +40,13 @@ if [ "$1" = "db" ]; then
         exit 0
     }
 
+    touch /DB_INITIALIZED
+
     trap stop_handler SIGTERM
 
     while true; do
-        sleep 1  # Sleep keeps CPU usage low while waiting
+        sleep 1 
     done
-
-    # child=$!
-    # wait "$child"
-
 fi
 
 echo "invalid command"
